requirements.txt
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
azure-functions==1.21.3
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.2.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
